Use getattr for oauth2_error access (#633) (#716)
* Use getattr for oauth2_error access (#633)

If the request doesn't have a oauth2_error property the
authenticate_header method errors. This can happen when the
oauthlib_core.verify_request method raises exceptions in authenticate.
It is useful to be able to raise AuthenticationFailed exceptions from
within a custom validate_bearer_token method which causes this.

* Add Test OAuth2Authentication authenticate override

Added a test for if an authenticate method that returns
None is used. This should result in a HTTP 401 response
for any request.